AI Engine/Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
bcrypt==4.0.1
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
python-jose==3.5.0
python-multipart==0.0.22
email-validator==2.1.1